Initial working version of get_virt script.

+++ get_virt.sh
@@ -27,10 +27,10 @@ echo "Installing dependencies..."
 
 case "$OS" in
 "arch")
-	pacman -S qemu libvirt edk2-ovmf virt-manager dnsmasq ebtables
+	yes | pacman -S qemu libvirt edk2-ovmf virt-manager dnsmasq ebtables swtpm
 	;;
 "pop" | "ubuntu" | "debian")
-	apt install qemu-kvm qemu-utils libvirt-daemon-system libvirt-clients bridge-utils virt-manager ovmf
+	apt install qemu-kvm qemu-utils libvirt-daemon-system libvirt-clients bridge-utils virt-manager ovmf swtpm swtpm-tools
 	;;
 *)
 	echo "$(tput setaf 2)Unsupported distribution: ${OS}.$(tput sgr0)"
@@ -38,6 +38,19 @@ case "$OS" in
 	;;
 esac
 
+echo "Adding kernel parameters."
+if command -v bootctl &>/dev/null; then
+	echo "Systemd-boot detected."
+	FILE=$(bootctl status | awk '/source:/ {print $2}')
+
+	cp $FILE "${FILE}.bak"
+	sed -i '/splash/ { /amd_iommu=on iommu=pt/! s/splash/amd_iommu=on iommu=pt splash/ }' $FILE
+else
+	echo "Only systemd-boot is supported. If you are running GRUB2 or something else, please add 'amd_iommu=on iommu=pt' to your kernel parameters."
+fi
+
+# TODO. Any operation needs to be trancendental aka. only run if operation would change something.
+echo "Enabling libvirtd."
 systemctl enable --now libvirtd
 virsh net-autostart default
 
@@ -45,12 +58,33 @@ echo "Adding user to groups."
 usermod -aG kvm,input,libvirt calin
 
 echo "Copying patched vgabios"
-mkdir -p /usr/share/vgabios/
-cp ./vbios/rtx3070-patched.rom /usr/share/vgabios/rtx3070-patched.rom
+if [ ! -f "/usr/share/vgabios/rtx3070-patched.rom" ]; then
+	mkdir -p /usr/share/vgabios/
+	cp ./vbios/rtx3070-patched.rom /usr/share/vgabios/rtx3070-patched.rom
+fi
 
 echo "Installing hooks."
-mkdir -p /etc/libvirt/hooks
-cp ./hooks/qemu /etc/libvirt/hooks/qemu
-chmod +x /etc/libvirt/hooks/qemu
+if [ ! -f "/etc/libvirt/hooks/qemu" ]; then
+	mkdir -p /etc/libvirt/hooks
+	cp ./hooks/qemu /etc/libvirt/hooks/qemu
+	chmod +x /etc/libvirt/hooks/qemu
+fi
 
-echo "Done! Restart before trying to launch your VM."
+if [ ! -f "/etc/libvirt/hooks/qemu.d/win11/prepare/begin/start.sh" ]; then
+	mkdir -p /etc/libvirt/hooks/qemu.d/win11/prepare/begin/
+	cp ./hooks/qemu.d/win11/prepare/begin/start.sh /etc/libvirt/hooks/qemu.d/win11/prepare/begin/start.sh
+	chmod +x /etc/libvirt/hooks/qemu.d/win11/prepare/begin/start.sh
+fi
+
+if [ ! -f "/etc/libvirt/hooks/qemu.d/win11/release/end/stop.sh" ]; then
+	mkdir -p /etc/libvirt/hooks/qemu.d/win11/release/end/
+	cp ./hooks/qemu.d/win11/release/end/stop.sh /etc/libvirt/hooks/qemu.d/win11/release/end/stop.sh
+	chmod +x /etc/libvirt/hooks/qemu.d/win11/release/end/stop.sh
+fi
+
+echo "Defining Windows 11 virtual machine."
+virsh define ./win11.xml
+
+echo "Done! A system restart is required.
+Also please make sure your Windows 11 image is placed at the following path:
+  $(tput bold)${HOME}/.libvirt/images/win11.qcow2$(tput sgr0)"
